functions mod

+++ src/models/clause.rs
@@ -3,7 +3,7 @@ use crate::models::lit::Lit;
 pub struct Clause {
     pub data: Vec<Lit>,
     is_learnt: bool,
-    activity: f32,
+    activity: f64,
 }
 
 pub trait IClause {
